pyproject.toml
[tool.poetry]
name = "perceptron-tkhm-alfai"
version = "1.0.0"
description = ""
authors = ["Muhammad Bilal Khan <lofi.audit@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
rich = "^13.8.1"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
openai = "^1.47.0"
pinecone-client = {extras = ["grpc"], version = "^5.0.1"}
langchain-community = "^0.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
